refactor(engine): move private class up
- also rename variables to be more descriptive
