+++ engine/src/main/java/io/camunda/zeebe/engine/processing/bpmn/container/MultiInstanceBodyProcessor.java
@@ -15,11 +15,11 @@ import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnEventSubscriptionBeh
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnIncidentBehavior;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnStateBehavior;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnStateTransitionBehavior;
+import io.camunda.zeebe.engine.processing.bpmn.behavior.OutputCollectionBehavior;
 import io.camunda.zeebe.engine.processing.common.ExpressionProcessor;
 import io.camunda.zeebe.engine.processing.common.Failure;
 import io.camunda.zeebe.engine.processing.deployment.model.element.ExecutableMultiInstanceBody;
 import io.camunda.zeebe.msgpack.spec.MsgPackHelper;
-import io.camunda.zeebe.msgpack.spec.MsgPackReader;
 import io.camunda.zeebe.msgpack.spec.MsgPackWriter;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.camunda.zeebe.protocol.record.value.ErrorType;
@@ -28,7 +28,6 @@ import io.camunda.zeebe.util.buffer.BufferUtil;
 import java.util.List;
 import java.util.Optional;
 import org.agrona.DirectBuffer;
-import org.agrona.ExpandableArrayBuffer;
 import org.agrona.MutableDirectBuffer;
 import org.agrona.concurrent.UnsafeBuffer;
 
@@ -349,106 +348,5 @@ public final class MultiInstanceBodyProcessor
           completionCondition.get(), context.getElementInstanceKey());
     }
     return Either.right(false);
-  }
-
-  private static final class OutputCollectionBehavior {
-
-    private final MsgPackReader variableReader = new MsgPackReader();
-    private final MsgPackWriter variableWriter = new MsgPackWriter();
-    private final ExpandableArrayBuffer variableBuffer = new ExpandableArrayBuffer();
-    private final DirectBuffer resultBuffer = new UnsafeBuffer(0, 0);
-
-    private final BpmnStateBehavior stateBehavior;
-    private final ExpressionProcessor expressionProcessor;
-
-    private OutputCollectionBehavior(
-        final BpmnStateBehavior stateBehavior, final ExpressionProcessor expressionProcessor) {
-      this.stateBehavior = stateBehavior;
-      this.expressionProcessor = expressionProcessor;
-    }
-
-    private void initializeOutputCollection(
-        final BpmnElementContext context, final DirectBuffer variableName, final int size) {
-
-      variableWriter.wrap(variableBuffer, 0);
-
-      // initialize the array with nil
-      variableWriter.writeArrayHeader(size);
-      for (var i = 0; i < size; i++) {
-        variableWriter.writeNil();
-      }
-
-      final var length = variableWriter.getOffset();
-
-      stateBehavior.setLocalVariable(context, variableName, variableBuffer, 0, length);
-    }
-
-    private Either<Failure, Void> updateOutputCollection(
-        final ExecutableMultiInstanceBody element,
-        final BpmnElementContext childContext,
-        final BpmnElementContext flowScopeContext) {
-
-      return element
-          .getLoopCharacteristics()
-          .getOutputCollection()
-          .map(
-              variableName ->
-                  updateOutputCollection(element, childContext, flowScopeContext, variableName))
-          .orElse(Either.right(null));
-    }
-
-    private Either<Failure, Void> updateOutputCollection(
-        final ExecutableMultiInstanceBody element,
-        final BpmnElementContext childContext,
-        final BpmnElementContext flowScopeContext,
-        final DirectBuffer variableName) {
-
-      final var loopCounter =
-          stateBehavior.getElementInstance(childContext).getMultiInstanceLoopCounter();
-
-      return readOutputElementVariable(element, childContext)
-          .map(
-              elementVariable -> {
-                // we need to read the output element variable before the current collection
-                // is read, because readOutputElementVariable(Context) uses the same
-                // buffer as getVariableLocal this could also be avoided by cloning the current
-                // collection, but that is slower.
-                final var currentCollection =
-                    stateBehavior.getLocalVariable(flowScopeContext, variableName);
-                final var updatedCollection =
-                    insertAt(currentCollection, loopCounter, elementVariable);
-                stateBehavior.setLocalVariable(flowScopeContext, variableName, updatedCollection);
-
-                return null;
-              });
-    }
-
-    private Either<Failure, DirectBuffer> readOutputElementVariable(
-        final ExecutableMultiInstanceBody element, final BpmnElementContext context) {
-      final var expression = element.getLoopCharacteristics().getOutputElement().orElseThrow();
-      return expressionProcessor.evaluateAnyExpression(expression, context.getElementInstanceKey());
-    }
-
-    private DirectBuffer insertAt(
-        final DirectBuffer array, final int index, final DirectBuffer element) {
-
-      variableReader.wrap(array, 0, array.capacity());
-      variableReader.readArrayHeader();
-      variableReader.skipValues((long) index - 1L);
-
-      final var offsetBefore = variableReader.getOffset();
-      variableReader.skipValue();
-      final var offsetAfter = variableReader.getOffset();
-
-      variableWriter.wrap(variableBuffer, 0);
-      variableWriter.writeRaw(array, 0, offsetBefore);
-      variableWriter.writeRaw(element);
-      variableWriter.writeRaw(array, offsetAfter, array.capacity() - offsetAfter);
-
-      final var length = variableWriter.getOffset();
-
-      resultBuffer.wrap(variableBuffer, 0, length);
-      return resultBuffer;
-    }
   }
 }
